requirements.txt
python-dotenv
langchain-openai
langgraph
reportlab